Farmer's Problem Price Effect
Problem: Assume the farmer allocates his land according the solution of 120 acres for wheat, 80 acres for corn, and 300 acres for sugar beets. When yields are good for the farmer, they are usually also good for many other farmers. The supply is thus increasing, which will lower the prices. As an example, we may consider prices going down by 10% for corn and wheat when yields are above average and going up by 10% when yields are below average. Formulate the model where these changes in prices affect both sales and purchases of corn and wheat. Assume that sugar beet prices are not affected by yields.

This file is the .dat file for the corresponding .mod file. These files run in AMPL