Farmer's Problem Binary First Stage
Consider the case where the farmer possesses four elds of sizes 185, 145,
105, and 65 acres, respectively. Observe that the total of 500 acres is unchanged. Now, the elds are
unfortunately located in dierent parts of the village. For reasons of eciency, the farmer wants
to race only one type of crop on each eld. Formulate this model as a two stage stochastic program
with a rst-stage program with binary variables.