Farmer's Problem - Risk Aversion
Economic theory tells us that, like many other people, the farmer would normally
act as a risk averse person. There are various ways to model risk aversion. One simple way is
to plan for the worst case. More precisely, it consists of maximizing the prot under the worst
situation. Note that for some models, it is not known in advance which scenario will turn out
to induce the lowest prot. In our example, the worst situation corresponds to Scenario 3 (below
average yields). Planning for the worst case implies that the solution of Table 4 is optimal.